Restructure folders and modules
