+++ src/index.tsx
@@ -1,6 +1,6 @@
-import * as Types from './types';
-export * from './FlashlessScript';
+export * from './components/FlashlessScript';
+export * from './providers/ColorModeToggleProvider';
 export * from './flashless';
-export * from './ColorModeToggle';
+import * as Types from './types';
 
 export {Types};
